update readme, break out cli into separate modules, add subcommands, cleanup, upgrade core
