+++ lib/unlock-schema.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = function(program) {
+  return {
+    properties: {
+      password: {
+        description: 'Unlock your private key to start storj',
+        hidden: true,
+        replace: '*',
+        required: true
+      }
+    }
+  };
+};
