+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "storjshare-cli",
-  "version": "3.0.6",
+  "version": "4.0.0",
   "description": "command line program for farming on the storj network",
   "preferGlobal": true,
   "bin": {
@@ -31,7 +31,7 @@
     "kad-logger-json": "^0.1.2",
     "myspeed": "^1.0.1",
     "prompt": "^1.0.0",
-    "storj": "^0.6.16",
+    "storj": "^1.0.1",
     "storj-telemetry-reporter": "Storj/telemetry-reporter"
   }
 }
